+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.2 - 2025-XX-XX
+
+### Added
+
+- Added checks to make sure there are no duplicate digests
+- Increased SD-JWT unpack strictness: now rejects on duplicate properties, invalid keys from disclosures, reserved key and invalid array digests.
+
+### Changed
+
+- Prevented use of reserved names (`_sd`, `...`) as keys for selectively disclosable claims.
+
 ## 1.2.1 - 2024-10-04
 
 ### Changed